tests: revert pr 1127
Reverting PR #1127 since Schema Registry tests were consistently failing after removal of deprecated `cgi.parse_header` usage:
```
aiohttp.client_exceptions.ContentTypeError: 500
Attempt to decode JSON with unexpected mimetype: text/plain; charset=utf-8
```
