+++ tests/e2e/schema_registry/test_jsonschema.py
@@ -5,11 +5,13 @@ See LICENSE for details
 
 import json
 
+import pytest
 from jsonschema import Draft7Validator
 
 from karapace.core.client import Client
 from karapace.core.compatibility import CompatibilityModes
 from karapace.core.schema_reader import SchemaType
+from karapace.core.typing import SchemaMetadata, SchemaRuleSet
 from tests.schemas.json_schemas import (
     A_DINT_B_DINT_OBJECT_SCHEMA,
     A_DINT_B_INT_OBJECT_SCHEMA,
@@ -23,6 +25,7 @@ from tests.schemas.json_schemas import (
     A_INT_OBJECT_SCHEMA,
     A_INT_OPEN_OBJECT_SCHEMA,
     A_OBJECT_SCHEMA,
+    ALL_SCHEMAS,
     ARRAY_OF_INT_SCHEMA,
     ARRAY_OF_NUMBER_SCHEMA,
     ARRAY_OF_POSITIVE_INTEGER,
@@ -230,6 +233,47 @@ async def not_schemas_are_backward_compatible(
         client=client,
         compatibility_mode=CompatibilityModes.BACKWARD,
     )
+
+
+@pytest.mark.parametrize("compatibility", [CompatibilityModes.FORWARD, CompatibilityModes.BACKWARD, CompatibilityModes.FULL])
+@pytest.mark.parametrize("metadata", [None, {}])
+@pytest.mark.parametrize("rule_set", [None, {}])
+async def test_same_jsonschema_must_have_same_id(
+    registry_async_client: Client,
+    compatibility: CompatibilityModes,
+    metadata: SchemaMetadata,
+    rule_set: SchemaRuleSet,
+) -> None:
+    for schema in ALL_SCHEMAS:
+        subject = new_random_name("subject")
+
+        res = await registry_async_client.put(f"config/{subject}", json={"compatibility": compatibility.value})
+        assert res.status_code == 200
+
+        first_res = await registry_async_client.post(
+            f"subjects/{subject}/versions",
+            json={
+                "schema": json.dumps(schema.schema),
+                "schemaType": SchemaType.JSONSCHEMA.value,
+                "metadata": metadata,
+                "ruleSet": rule_set,
+            },
+        )
+        assert first_res.status_code == 200
+        first_id = first_res.json().get("id")
+        assert first_id
+
+        second_res = await registry_async_client.post(
+            f"subjects/{subject}/versions",
+            json={
+                "schema": json.dumps(schema.schema),
+                "schemaType": SchemaType.JSONSCHEMA.value,
+                "metadata": metadata,
+                "ruleSet": rule_set,
+            },
+        )
+        assert second_res.status_code == 200
+        assert first_id == second_res.json()["id"]
 
 
 async def test_schemaregistry_schemaregistry_extra_optional_field_with_open_model_is_compatible(
